# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Invoice Exception Resolution \u2014 Dashboard & KB API",
    "description": "RESTful API for the autonomous LangGraph agent: dashboard search, human approval, knowledge-base search, and analytics. The six-gate decision pipeline runs asynchronously via Celery.",
    "version": "4.0.0"
  },
  "paths": {
...
    "/tools/approve/{exception_id}": {
      "post": {
        "tags": [
          "Human Approval"
        ],
        "summary": "Manually Approve Escalated Exception",
        "operationId": "approve_tools_approve__exception_id__post",
        "parameters": [
          {
            "name": "exception_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exception Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/reject/{exception_id}": {
      "post": {
        "tags": [
          "Human Approval"
        ],
        "summary": "Manually Reject Escalated Exception",
        "operationId": "reject_tools_reject__exception_id__post",
        "parameters": [
          {
            "name": "exception_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exception Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exceptions/list": {
      "post": {
        "tags": [
          "Dashboard"
        ],
        "summary": "List Exceptions with Search/Filter",
        "operationId": "list_exceptions_exceptions_list_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExceptionListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/kb/search/emails": {
      "post": {
        "tags": [
          "Knowledge Base"
        ],
        "summary": "Semantic Email Search",
        "operationId": "search_emails_kb_search_emails_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KBEmailSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KBEmailSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/kb/search/transcripts": {
      "post": {
        "tags": [
          "Knowledge Base"
        ],
        "summary": "Semantic Transcript Search",
        "operationId": "search_transcripts_kb_search_transcripts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KBTranscriptSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KBTranscriptSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/kb/history/{supplier_id}": {
      "get": {
        "tags": [
          "Knowledge Base"
        ],
        "summary": "Supplier Resolution History",
        "operationId": "supplier_history_kb_history__supplier_id__get",
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Supplier Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Supplier History Kb History  Supplier Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/summary": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get KPI Dashboard Summary",
        "operationId": "get_analytics_summary_analytics_summary_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/parse": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Parse a PDF document via OCR + LLM",
        "description": "Upload a PDF (invoice, PO, or GRN) and receive the parsed Pydantic model. OCR uses Tesseract; structured extraction uses the configured OpenAI-compatible LLM. The endpoint does NOT auto-create an exception \u2014 review the parsed model, then call /webhook/invoice, /webhook/po, or /webhook/grn to enqueue it.",
        "operationId": "parse_document_documents_parse_post",
        "parameters": [
          {
            "name": "doc_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Type"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_document_documents_parse_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentParseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "properties": {
          "approved_by": {
            "type": "string",
            "title": "Approved By",
            "description": "User ID or email of approver"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Optional approval notes"
          }
        },
        "type": "object",
        "required": [
          "approved_by"
        ],
        "title": "ApprovalRequest"
      },
      "ApprovalResponse": {
        "properties": {
          "exception_id": {
            "type": "string",
            "title": "Exception Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "exception_id",
          "status",
          "message"
        ],
        "title": "ApprovalResponse"
      },
...
      "ExceptionListRequest": {
        "properties": {
          "supplier_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "supplier_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Name"
          },
          "invoice_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Number"
          },
          "po_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Po Number"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "variance_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variance Min"
          },
          "variance_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variance Max"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 50
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "title": "ExceptionListRequest"
      },
      "ExceptionListResponse": {
        "properties": {
          "exceptions": {
            "items": {
              "$ref": "#/components/schemas/ExceptionSummary"
            },
            "type": "array",
            "title": "Exceptions"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "exceptions",
          "total_count",
          "limit",
          "offset"
        ],
        "title": "ExceptionListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KBEmailSearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Results"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "query",
          "results",
          "total"
        ],
        "title": "KBEmailSearchResponse"
      },
...
      "RejectionRequest": {
        "properties": {
          "rejected_by": {
            "type": "string",
            "title": "Rejected By",
            "description": "User ID or email of reviewer"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Reason for rejection"
          }
        },
        "type": "object",
        "required": [
          "rejected_by",
          "reason"
        ],
        "title": "RejectionRequest"
      },
      "RejectionResponse": {
        "properties": {
          "exception_id": {
            "type": "string",
            "title": "Exception Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "exception_id",
          "status",
          "message"
        ],
        "title": "RejectionResponse"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      }
    }
  }
}